pyproject.toml
[project]
name = "auto-trader"
version = "0.1.0"
description = "Paper-first AI trading bot foundation inspired by proven open-source frameworks."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alpaca-py>=0.42.0",
    "joblib>=1.4.2",
    "numpy>=2.2.2",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.1",
    "yfinance>=0.2.61",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
]

[project.scripts]
autotrader = "src.cli:main"

[build-system]
requires = ["setuptools>=75.8.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-q"
